# packages/api-client/openapi.yaml
openapi: 3.0.3
info:
  title: Boccone AI API
  version: 0.1.0
  description: Authenticated application API for Boccone AI.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
  /api/health:
    get:
      operationId: getHealth
      tags:
        - health
      summary: Check API health
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "500":
          $ref: "#/components/responses/ErrorResponse"
  /api/me:
    get:
      operationId: getCurrentUser
      tags:
        - session
      summary: Get the current authenticated user
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/users:
    get:
      operationId: listAdminUsers
      tags:
        - admin
      summary: List users for administrators
      security:
        - sessionCookie: []
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 255
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Users visible to operations staff
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createAdminUser
      tags:
        - admin
      summary: Create a user for administrators
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserCreateRequest"
      responses:
        "200":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getAdminUser
      tags:
        - admin
      summary: Get a user for administrators
      security:
        - sessionCookie: []
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      operationId: updateAdminUser
      tags:
        - admin
      summary: Update a user for administrators
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdateRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: removeAdminUser
      tags:
        - admin
      summary: Remove a user for administrators
      security:
        - sessionCookie: []
      responses:
        "200":
          description: User removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMutationResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/users/{id}/role:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: setAdminUserRole
      tags:
        - admin
      summary: Change a user role
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserRoleRequest"
      responses:
        "200":
          description: Role-updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/users/{id}/ban:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: banAdminUser
      tags:
        - admin
      summary: Ban a user
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserBanRequest"
      responses:
        "200":
          description: Banned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/users/{id}/unban:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: unbanAdminUser
      tags:
        - admin
      summary: Unban a user
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Unbanned user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/admin/audit-logs:
    get:
      operationId: listAdminAuditLogs
      tags:
        - admin
      summary: List admin audit logs
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Admin audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuditLogsResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "403":
          $ref: "#/components/responses/ErrorResponse"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  responses:
    ErrorResponse:
      description: API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - bad_request
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - validation_error
                - internal_error
            message:
              type: string
            requestId:
              type: string
    HealthResponse:
      type: object
      required:
        - status
        - service
        - version
        - requestId
        - timestamp
      properties:
        status:
          type: string
          enum: [ok]
        service:
          type: string
          enum: [boccone-api]
        version:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
    AuthUser:
      type: object
      required:
        - id
        - name
        - email
        - emailVerified
        - image
        - role
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        image:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
    MeResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
    AdminUser:
      type: object
      required:
        - id
        - name
        - email
        - emailVerified
        - image
        - role
        - banned
        - banReason
        - banExpires
        - createdAt
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        image:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        banned:
          type: boolean
          nullable: true
        banReason:
          type: string
          nullable: true
        banExpires:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    AdminUsersResponse:
      type: object
      required:
        - users
        - total
        - limit
        - offset
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    AdminUserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
    AdminUserCreateRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        role:
          type: string
          enum: [user, admin]
          default: user
    AdminUserUpdateRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        email:
          type: string
          format: email
    AdminUserRoleRequest:
      type: object
      required:
        - role
      properties:
        role:
          type: string
          enum: [user, admin]
    AdminUserBanRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
        durationSeconds:
          type: integer
          minimum: 60
          maximum: 31536000
    AdminMutationResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          enum: [true]
    AdminAuditAction:
      type: string
      enum:
        - user_created
        - user_updated
        - user_role_changed
        - user_banned
        - user_unbanned
        - user_removed
    AdminAuditLog:
      type: object
      required:
        - id
        - actorUserId
        - targetUserId
        - action
        - metadata
        - createdAt
      properties:
        id:
          type: string
        actorUserId:
          type: string
        targetUserId:
          type: string
          nullable: true
        action:
          $ref: "#/components/schemas/AdminAuditAction"
        metadata:
          type: object
          additionalProperties:
            oneOf:
              - type: string
              - type: number
              - type: boolean
        createdAt:
          type: string
          format: date-time
    AdminAuditLogsResponse:
      type: object
      required:
        - logs
        - total
        - limit
        - offset
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AdminAuditLog"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer